--- extension/blocked.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <body>
-        <h1>This request was blocked for security</h1>
-        <p>
-            <b>Error:</b> <span id="msg"></span><br />
-            <b>Failed URL:</b> <a href="" id="url"></a>
-        </p>
-        <script type="text/javascript" src="blocked.js"></script>
-    </body>
-</html>
