--- extension/popup.html
@@ -1,17 +1,0 @@
-<!doctype html>
-
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <link rel="stylesheet" href="popup.css"/>
-    <script type="text/javascript" src="jquery.js"></script>
-    <title>Attested DNS Status Information</title>
-  </head>
-
-  <body>
-    <h1>aDNS Info: <span id="name"></span></h1>
-    <table id="tab"></table>
-
-    <script src="popup.js"></script>
-  </body>
-</html>
