+++ include/adns_types.h
