+++ include/resolver.h
