--- key.conf
@@ -1,4 +1,0 @@
-key "karan" {
-        algorithm hmac-sha256;
-        secret "Hiu9edsm7Punk9jGaSF7VYoFRzqzcdOKrxv7u3nPf6w=";
-};
