--- oe_sign.conf
@@ -1,7 +1,0 @@
-# Enclave settings:
-NumHeapPages=50000
-NumStackPages=1024
-NumTCS=8
-ProductID=1
-SecurityVersion=1
-# The Debug setting is automatically inserted by sign_app_library in CMake, to build both debuggable and non-debuggable variants
