+++ src/base32.cpp
